refactor(formats): split gene-model parser handlers
